Refactor TransactionProcessor.run() loop
This commit removes the long blocking behavior of the run() loop on
TransactionProcessor.
Add getReceiveQueue() method to Stream and ZmqStream, so that we can
poll without TimeOutException
Add synchronous behavior around the handlers collection.
Move MessageWrapper out of SendReceiveThread since it is now used
independently.
Refactor TransactionProcessor to create the handleMessage method so that
unit tests can be made deterministic
Updated unit test to be deterministic.

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>